requirements.txt
#add together ai
termcolor
pandas
pexpect
scipy
ipdb
absl-py
diskcache
python-dotenv